requirements.txt
# Flask Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Vector Database
chromadb==0.5.23

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Groq API Client
groq==0.9.0

# OpenAI API Client
openai==1.12.0

# Environment & Utilities
python-dotenv==1.0.0
requests==2.31.0

# NumPy - Compatible version
numpy==1.26.4

# Additional dependencies
httpx==0.27.0
pydantic==2.6.0